cleartmp {summarytools}	R Documentation

Delete Temporary Html Files

Description

Delete temporary files created when using generic print method with method='browser' or method='viewer', or when calling view() function.

Usage

cleartmp(all = TRUE, silent = FALSE, verbose = FALSE)

Arguments

all
Logical. When TRUE (default), all temporary summarytools files are deleted. When FALSE, only the latest file is.

silent
Logical. Hide confirmation messages (FALSE by default).

verbose
Logical. Display a message for every file that is deleted. FALSE by default.

Note

Given that all temporary files are deleted automatically when an R session is ended, this function is an overkill in most circumstances. It could however be useful in server-type setups.

Author(s)

Dominic Comtois, dominic.comtois@gmail.com

[Package summarytools version 1.0.1 Index]