Univariate Statistics for Numerical Data

Calculates mean, sd, min, Q1*, median, Q3*, max, MAD, IQR*, CV, skewness*, SE.skewness*, and kurtosis* on numerical vectors. (*) Not available when using sampling weights.

descr(x, stats = st_options("descr.stats"), na.rm = TRUE,
  round.digits = st_options("round.digits"),
  transpose = st_options("descr.transpose"),
  style = st_options("style"), plain.ascii = st_options("plain.ascii"),
  justify = "right", omit.headings = st_options("omit.headings"),
  display.labels = st_options("display.labels"), split.tables = 100,
  weights = NA, rescale.weights = FALSE, ...)

Arguments

x
A numerical vector or a data frame.

stats
Which stats to produce. Either “all” (default), or a selection of : “mean”, “sd”, “min”, “q1”, “med”, “q3”, “max”, “mad”, “iqr”, “cv”, “skewness”, “se.skewness”, “kurtosis”, “n.valid”, and “pct.valid”. This can be set globally via st_options (“descr.stats”).

na.rm
Argument to be passed to statistical functions. Defaults to TRUE. Can be set globally; see st_options.

round.digits
Number of significant digits to display. Defaults to 2, and can be set globally (see st_options).

transpose
Logical. Makes variables appears as columns, and stats as rows. Defaults to FALSE. To change this default value, see st_options (option “descr.transpose”).

style
Style to be used by pander when rendering output table; One of “simple” (default), “grid”, or “rmarkdown” This option can be set globally; see st_options.

plain.ascii
Logical. pander argument; when TRUE, no markup characters will be used (useful when printing to console). Defaults to TRUE unless style = 'rmarkdown', in which case it will be set to FALSE automatically. To change the default value globally, see st_options.

justify
Alignment of numbers in cells; “l” for left, “c” for center, or “r” for right (default). Has no effect on html tables.

omit.headings
Logical. Set to TRUE to omit heading section. Can be set globally via st_options.

display.labels
Logical. Should variable / data frame labels be displayed in the title section? Default is TRUE. To change this default value globally, see st_options.

split.tables
Pander argument that specifies how many characters wide a table can be. 100 by default.

weights
Vector of weights having same length as x. NA (default) indicates that no weights are used.

rescale.weights
Logical. When set to TRUE, the total count will be the same as the unweighted x. FALSE by default.

…
Additional arguments passed to pander.

Value

A nn object of classes matrix and summarytools containing the statistics, with extra attributes used by print method.

Examples

data(exams)
descr(exams)

#> Non-numerical variable(s) ignored: student, gender

#> Descriptive Statistics     
#> Data Frame: exams     
#> N: 30   
#> 
#>                     french    math   geography   history   economics   english
#> ----------------- -------- ------- ----------- --------- ----------- ---------
#>              Mean    73.94   73.54       70.04     72.77       73.91     75.96
#>           Std.Dev    10.79    9.19       10.65     10.20        8.62      7.92
#>               Min    44.80   55.60       47.20     53.90       60.50     58.30
#>                Q1    68.20   66.95       65.90     68.20       68.80     70.90
#>            Median    73.60   73.75       68.50     72.75       71.60     74.10
#>                Q3    76.70   80.35       77.80     76.50       77.00     80.60
#>               Max    94.70   93.20       96.30     93.50       94.20     93.10
#>               MAD     7.56    9.93       12.31      6.45        5.49      6.52
#>               IQR     8.50   13.35       11.90      8.15        8.20      9.70
#>                CV     0.15    0.12        0.15      0.14        0.12      0.10
#>          Skewness     0.03    0.12        0.10      0.01        0.75      0.28
#>       SE.Skewness     0.43    0.44        0.43      0.43        0.43      0.43
#>          Kurtosis     0.45   -0.58       -0.03     -0.60       -0.42     -0.25
#>           N.Valid    29.00   28.00       29.00     30.00       29.00     29.00
#>         Pct.Valid    96.67   93.33       96.67    100.00       96.67     96.67

descr(exams, stats = c("mean", "sd", "min", "max"), transpose = TRUE)

#> Non-numerical variable(s) ignored: student, gender

#> Descriptive Statistics     
#> Data Frame: exams     
#> N: 30   
#> 
#>                    Mean   Std.Dev     Min     Max
#> --------------- ------- --------- ------- -------
#>          french   73.94     10.79   44.80   94.70
#>            math   73.54      9.19   55.60   93.20
#>       geography   70.04     10.65   47.20   96.30
#>         history   72.77     10.20   53.90   93.50
#>       economics   73.91      8.62   60.50   94.20
#>         english   75.96      7.92   58.30   93.10

data(tobacco)
with(tobacco, view(by(BMI, gender, descr), method = "pander"))

#> Descriptive Statistics   
#> tobacco$BMI by gender   
#> 
#>                          F        M
#> ----------------- -------- --------
#>              Mean    26.10    25.31
#>           Std.Dev     4.95     3.98
#>               Min     9.01     8.83
#>                Q1    22.98    22.52
#>            Median    25.87    25.14
#>                Q3    29.48    27.96
#>               Max    39.44    36.76
#>               MAD     4.75     4.02
#>               IQR     6.49     5.44
#>                CV     0.19     0.16
#>          Skewness    -0.02    -0.04
#>       SE.Skewness     0.11     0.11
#>          Kurtosis     0.09     0.17
#>           N.Valid   475.00   477.00
#>         Pct.Valid    97.14    97.55