Data frame Summary

Summary of a data frame consisting of: variable names, labels if any, factor levels, frequencies and/or numerical summary statistics, and valid/missing observation counts.

dfSummary(x, round.digits = st_options("round.digits"),
  varnumbers = st_options("dfSummary.varnumbers"),
  labels.col = length(label(x, all = TRUE)) > 0,
  valid.col = st_options("dfSummary.valid.col"),
  na.col = st_options("dfSummary.na.col"),
  graph.col = st_options("dfSummary.graph.col"),
  graph.magnif = st_options("dfSummary.graph.magnif"),
  style = "multiline", plain.ascii = st_options("plain.ascii"),
  justify = "left", omit.headings = st_options("omit.headings"),
  max.distinct.values = 10, trim.strings = FALSE,
  max.string.width = 25, split.cells = 40, split.tables = Inf, ...)

Arguments

x
A data frame.

round.digits
Number of significant digits to display. Defaults to 2 and can be set globally; see st_options.

varnumbers
Logical. Should the first column contain variable number? Defaults to TRUE. Can be set globally; see st_options, option “dfSummary.varnumbers”.

labels.col
Logical. If TRUE, variable labels (as defined with rapportools, Hmisc or summarytools' label functions) will be displayed. By default, the labels column is shown if at least one column has a defined label.

valid.col
Logical. Include column indicating count and proportion of valid (non-missing) values. TRUE by default, but can be set globally; see st_options, option “dfSummary.valid.col”.

na.col
Logical. Include column indicating count and proportion of missing (NA) values. TRUE by default, but can be set globally; see st_options, option “dfSummary.na.col”.

graph.col
Logical. Display barplots / histograms column in html reports. TRUE by default, but can be set globally; see st_options, option “dfSummary.graph.col”.

graph.magnif
Numeric. Magnification factor, useful if the graphs show up too large (then use a value < 1) or too small (use a value > 1). Must be positive. Can be set globally; see st_options, option “dfSummary.graph.magnif”.

style
Style to be used by pander when rendering output table. Defaults to “multiline”. The only other valid option is “grid”. Style “simple” is not supported for this particular function, and “rmarkdown” will fallback to “multiline”.

plain.ascii
Logical. pander argument; when TRUE, no markup characters will be used (useful when printing to console). Defaults to TRUE. Set to FALSE when in context of markdown rendering. To change the default value globally, see st_options.

justify
String indicating alignment of columns; one of “l” (left) “c” (center), or “r” (right). Defaults to “l”.

omit.headings
Logical. Set to TRUE to omit headings. To change this default value globally, see st_options.

max.distinct.values
The maximum number of values to display frequencies for. If variable has more distinct values than this number, the remaining frequencies will be reported as a whole, along with the number of additional distinct values. Defaults to 10.

trim.strings
Logical; for character variables, should leading and trailing white space be removed? Defaults to FALSE. See details section.

max.string.width
Limits the number of characters to display in the frequency tables. Defaults to 25.

split.cells
A numeric argument passed to pander. It is the number of characters allowed on a line before splitting the cell. Defaults to 40.

split.tables
pander argument which determines the maximum width of a table. Keeping the default value (Inf) is recommended.

…
Additional arguments passed to pander.

Value

A data frame with additional class summarytools containing as many rows as there are columns in x, with attributes to inform print method. Columns in the output data frame are:

No

Number indicating the order in which column appears in the data frame.

Variable

Name of the variable, along with its class(es).

Label

Label of the variable (if applicable).

Stats / Values

For factors, a list of their values, limited by the max.distinct.values parameter. For character variables, the most common values (in descending frequency order), also limited by max.distinct.values. For numerical variables, common univariate statistics (mean, std. deviation, min, med, max, IQR and CV).

Freqs (% of Valid)

For factors and character variables, the frequencies and proportions of the values listed in the previous column. For numerical vectors, number of distinct values, or frequency of distinct values if their number is not greater than max.distinct.values.

Text Graph

An ascii histogram for numerical variables, and ascii barplot for factors and character variables.

Valid

Number and proportion of valid values.

Missing

Number and proportion of missing (NA and NAN) values.

Details

The default plain.ascii = TRUE option is there to make results appear cleaner in the console. When used in a context of rmarkdown rendering, set this option to FALSE.

When the trim.strings is set to TRUE, trimming is done before calculating frequencies, so those will be impacted accordingly.

The package vignette “Recommendations for Rmarkdown” provides valuable information for creating optimal Rmarkdown documents with summarytools.

Examples

data(tobacco)
dfSummary(tobacco)

#> Data Frame Summary   
#> tobacco     
#> N: 1000     
#> Duplicates: 2   
#> -------------------------------------------------------------------------------------------------------------------------
#> No   Variable        Stats / Values              Freqs (% of Valid)    Text Graph                     Valid     Missing  
#> ---- --------------- --------------------------- --------------------- ------------------------------ --------- ---------
#> 1    gender          1. F                        489 (50.0%)           IIIIIIIIIIIIIIII               978       22       
#>      [factor]        2. M                        489 (50.0%)           IIIIIIIIIIIIIIII               (97.8%)   (2.2%)   
#> 
#> 2    age             mean (sd) : 49.6 (18.29)    63 distinct values    .     .     . . . :            975       25       
#>      [numeric]       min < med < max :                                 : : : : : . : : : :            (97.5%)   (2.5%)   
#>                      18 < 50 < 80                                      : : : : : : : : : :                               
#>                      IQR (CV) : 32 (0.37)                              : : : : : : : : : :                               
#>                                                                        : : : : : : : : : :                               
#> 
#> 3    age.gr          1. 18-34                    258 (26.5%)           IIIIIIIIIIIII                  975       25       
#>      [factor]        2. 35-50                    241 (24.7%)           IIIIIIIIIIII                   (97.5%)   (2.5%)   
#>                      3. 51-70                    317 (32.5%)           IIIIIIIIIIIIIIII                                  
#>                      4. 71 +                     159 (16.3%)           IIIIIIII                                          
#> 
#> 4    BMI             mean (sd) : 25.73 (4.49)    974 distinct values             :                    974       26       
#>      [numeric]       min < med < max :                                         : : :                  (97.4%)   (2.6%)   
#>                      8.83 < 25.62 < 39.44                                      : : :                                     
#>                      IQR (CV) : 5.72 (0.17)                                  : : : : :                                   
#>                                                                            . : : : : : .                                 
#> 
#> 5    smoker          1. Yes                      298 (29.8%)           IIIIII                         1000      0        
#>      [factor]        2. No                       702 (70.2%)           IIIIIIIIIIIIIIII               (100%)    (0%)     
#> 
#> 6    cigs.per.day    mean (sd) : 6.78 (11.88)    37 distinct values    :                              965       35       
#>      [numeric]       min < med < max :                                 :                              (96.5%)   (3.5%)   
#>                      0 < 0 < 40                                        :                                                 
#>                      IQR (CV) : 11 (1.75)                              :                                                 
#>                                                                        :   . . . . . .                                   
#> 
#> 7    diseased        1. Yes                      224 (22.4%)           IIII                           1000      0        
#>      [factor]        2. No                       776 (77.6%)           IIIIIIIIIIIIIIII               (100%)    (0%)     
#> 
#> 8    disease         1. Hypertension             36 (16.2%)            IIIIIIIIIIIIIIII               222       778      
#>      [character]     2. Cancer                   34 (15.3%)            IIIIIIIIIIIIIII                (22.2%)   (77.8%)  
#>                      3. Cholesterol              21 ( 9.5%)            IIIIIIIII                                         
#>                      4. Heart                    20 ( 9.0%)            IIIIIIII                                          
#>                      5. Pulmonary                20 ( 9.0%)            IIIIIIII                                          
#>                      6. Musculoskeletal          19 ( 8.6%)            IIIIIIII                                          
#>                      7. Diabetes                 14 ( 6.3%)            IIIIII                                            
#>                      8. Hearing                  14 ( 6.3%)            IIIIII                                            
#>                      9. Digestive                12 ( 5.4%)            IIIII                                             
#>                      10. Hypotension             11 ( 5.0%)            IIII                                              
#>                      [ 3 others ]                21 ( 9.5%)            IIIIIIIII                                         
#> 
#> 9    samp.wgts       mean (sd) : 1 (0.08)        0.86!: 267 (26.7%)    IIIIIIIIIIIII                  1000      0        
#>      [numeric]       min < med < max :           1.04!: 249 (24.9%)    IIIIIIIIIIII                   (100%)    (0%)     
#>                      0.86 < 1.04 < 1.06          1.05!: 324 (32.4%)    IIIIIIIIIIIIIIII                                  
#>                      IQR (CV) : 0.19 (0.08)      1.06!: 160 (16.0%)    IIIIIII                                           
#>                                                  ! rounded                                                               
#> -------------------------------------------------------------------------------------------------------------------------
# NOT RUN {
view(dfSummary(iris))
# }