examens {summarytools}	R Documentation

Bulletin de notes (donnees simulees)

Description

Jeu de donnees simulees contenant les notes de 30 etudiants, avec les colonnes suivantes:

etudiant Nom de l'etudiant.

sexe Variable categorielle (facteur). Deux niveaux: “Fille”, “Garcon”.

francais Note en francais (numerique).

math Note en maths (numerique).

geographie Note en geographie (numerique).

histoire Note en histoire (numerique).

economie Note en economie (numerique).

anglais Note en anglais (numerique).

Usage

data(examens)

Format

Un data frame de 30 rangees et 8 colonnes

Details

Donnees simulees. Les notes de chaque etudiant sont centrees autour d'une moyenne personnelle et ecart-type randomises.

A copy of this dataset is available in English under the name “exams”.

[Package summarytools version 1.0.1 Index]