exams {summarytools}	R Documentation

Report Cards - Simulated Data

Description

A simulated dataset with grades for hypothetical 30 students, with the following variables:

student Student's name.

gender Factor with 2 levels: “Girl”, “Boy”.

french French Grade (numerical).

math Math Grade (numerical).

geography Geography Grade (numerical).

history History Grade (numerical).

economics Economics Grade (numerical).

english English Grade (numerical).

Usage

data(exams)

Format

A data frame with 30 rows and 8 variables

Details

All names and grades are simulated. Grades for each student are centered around a personal randomized average and standard deviation.

A copy of this dataset is also available in French under the name “examens”.

[Package summarytools version 1.0.1 Index]