Frequency Tables for Factors and Other Discrete Data

Displays weighted or unweighted frequencies, including <NA> counts and proportions.

freq(x, round.digits = st_options("round.digits"), order = "names",
  style = st_options("style"), plain.ascii = st_options("plain.ascii"),
  justify = "default", totals = st_options("freq.totals"),
  report.nas = st_options("freq.report.nas"), missing = "",
  display.type = TRUE, display.labels = st_options("display.labels"),
  omit.headings = st_options("omit.headings"), weights = NA,
  rescale.weights = FALSE, ...)

Arguments

x
Factor or vector

round.digits
Number of significant digits to display. Defaults to 2 and can be set globally; see st_options.

order
Ordering of rows in frequency table; “names” (default for non-factors), “levels” (default for factors), or “freq” (from most frequent to less frequent).

style
Style to be used by pander when rendering output table; One of “simple” (default), “grid”, or “rmarkdown” This option can be set globally; see st_options.

plain.ascii
Logical. pander argument; when TRUE, no markup characters will be used (useful when printing to console). Defaults to TRUE unless style = 'rmarkdown', in which case it will be set to FALSE automatically. To change the default value globally, see st_options.

justify
String indicating alignment of columns. By default (“default”), “right” is used for text tables and “center” is used for html tables. You can force it to one of “left”, “center”, or “right”.

totals
Logical. Set to FALSE to hide totals from results. To change this value globally, see st_options.

report.nas
Logical. Set to FALSE to turn off reporting of missing values. To change this default value globally, see st_options.

missing
Characters to display in NA cells. Defaults to “”.

display.type
Logical. Should variable type be displayed? Default is TRUE.

display.labels
Logical. Should variable / data frame labels be displayed? Default is TRUE. To change this default value globally, see st_options.

omit.headings
Logical. Set to TRUE to omit heading section. Can be set globally via st_options.

weights
Vector of weights; must be of the same length as x.

rescale.weights
Logical parameter. When set to TRUE, the total count will be the same as the unweighted x. FALSE by default.

…
Additional arguments passed to pander.

Value

A frequency table of class matrix and summarytools with added attributes used by print method.

Details

The default plain.ascii = TRUE option is there to make results appear cleaner in the console. To avoid rmarkdown rendering problems, this option is automatically set to FALSE whenever style = "rmarkdown" (unless plain.ascii = TRUE is made explicit in the function call).

See also

Examples

data(tobacco)
freq(tobacco$gender)

#> Frequencies   
#> tobacco$gender     
#> Type: Factor (unordered)   
#> 
#>               Freq   % Valid   % Valid Cum.   % Total   % Total Cum.
#> ----------- ------ --------- -------------- --------- --------------
#>           F    489     50.00          50.00     48.90          48.90
#>           M    489     50.00         100.00     48.90          97.80
#>        <NA>     22                               2.20         100.00
#>       Total   1000    100.00         100.00    100.00         100.00

freq(tobacco$gender, totals = FALSE)

#> Frequencies   
#> tobacco$gender     
#> Type: Factor (unordered)   
#> 
#>              Freq   % Valid   % Valid Cum.   % Total   % Total Cum.
#> ---------- ------ --------- -------------- --------- --------------
#>          F    489     50.00          50.00     48.90          48.90
#>          M    489     50.00         100.00     48.90          97.80
#>       <NA>     22                               2.20         100.00

freq(tobacco$gender, display.nas = FALSE)

#> Frequencies   
#> tobacco$gender     
#> Type: Factor (unordered)   
#> 
#>               Freq   % Valid   % Valid Cum.   % Total   % Total Cum.
#> ----------- ------ --------- -------------- --------- --------------
#>           F    489     50.00          50.00     48.90          48.90
#>           M    489     50.00         100.00     48.90          97.80
#>        <NA>     22                               2.20         100.00
#>       Total   1000    100.00         100.00    100.00         100.00

freq(tobacco$gender, style="rmarkdown")

#> ### Frequencies   
#> **Variable:** tobacco$gender     
#> **Type:** Factor (unordered)   
#> 
#> |     &nbsp; | Freq | % Valid | % Valid Cum. | % Total | % Total Cum. |
#> |-----------:|-----:|--------:|-------------:|--------:|-------------:|
#> |      **F** |  489 |   50.00 |        50.00 |   48.90 |        48.90 |
#> |      **M** |  489 |   50.00 |       100.00 |   48.90 |        97.80 |
#> | **\<NA\>** |   22 |         |              |    2.20 |       100.00 |
#> |  **Total** | 1000 |  100.00 |       100.00 |  100.00 |       100.00 |

with(tobacco, view(by(diseased, smoker, freq), method = "pander"))

#> Frequencies   
#> tobacco$diseased     
#> Type: Factor (unordered)     
#> Group: smoker = Yes   
#> 
#>               Freq   % Valid   % Valid Cum.   % Total   % Total Cum.
#> ----------- ------ --------- -------------- --------- --------------
#>         Yes    125     41.95          41.95     41.95          41.95
#>          No    173     58.05         100.00     58.05         100.00
#>        <NA>      0                               0.00         100.00
#>       Total    298    100.00         100.00    100.00         100.00
#>   
#> Group: smoker = No   
#> 
#>               Freq   % Valid   % Valid Cum.   % Total   % Total Cum.
#> ----------- ------ --------- -------------- --------- --------------
#>         Yes     99     14.10          14.10     14.10          14.10
#>          No    603     85.90         100.00     85.90         100.00
#>        <NA>      0                               0.00         100.00
#>       Total    702    100.00         100.00    100.00         100.00