label {summarytools}	R Documentation

Get or Set Variable or Data Frame Labels

Description

Assigns a label to a vector or data frame, or returns value stored in the object's label attribute (or NA if none exists).

Usage

label(x, all = FALSE, fallback = FALSE, simplify = FALSE)
label(x) <- value

Arguments

x
An R object to extract labels from.

all
Logical. When x is a data frame, setting this argument to TRUE will make the function return all variable labels. By default, its value is FALSE, so that if x is a data frame, it is the data frame's label that will be returned.

fallback
a logical value indicating if labels should fallback to object name(s). Defaults to FALSE.

simplify
When x is a data frame and all = TRUE, coerce results to a vector and remove NA's. Default is FALSE.

value
String to be used as label. To clear existing labels, use NA or NULL.

Note

Loosely based on Gergely Daróczi's label function.

Author(s)

Dominic Comtois, dominic.comtois@gmail.com,

[Package summarytools version 1.0.1 Index]