Extract Data Information From Arguments Passed to Functions

Using sys.calls(), sys.frames() and match.call(), this utility function extracts and/or deducts information about the data being processed. Data frame name, variable names and labels if any, subsetting information, grouping information (when by() is used) are returned by the function which tries various methods to get this information.

parse_args(sys_calls, sys_frames, match_call, var = "x")

Arguments

sys_calls
Object created using sys.calls().

sys_frames
Object created using sys.frames().

match_call
Object created using match.call().

var
One of “x” or “y” (the latter being used only in ctable).

Value

A list consisting of:

df_name The data frame name when applicable.

var_names The variable names when applicable.

rows_subset The subsetting condition when applicable.

by_group The group, when functions are called through by()

by_first A binary indicator used when function was called through by()

by_last A binary indicator used when function was called through by()