Print Method for Objects of Class summarytools.

Display summarytools objects in the console, in Web Browser or in RStudio's Viewer, or write content to file.

# S3 method for summarytools
print(x, method = "pander", file = "",
   append = FALSE, report.title = NA, table.classes = NA,
   bootstrap.css = st_options('bootstrap.css'), 
   custom.css = st_options('custom.css'), silent = FALSE, 
   footnote = st_options('footnote'), 
   escape.pipe = st_options('escape.pipe'), …)

view(x, method = "viewer", file = "", append = FALSE,
  report.title = NA, table.classes = NA, 
  bootstrap.css = st_options('bootstrap.css'), 
  custom.css = st_options('custom.css'), silent = FALSE, 
  footnote = st_options('footnote'), 
  escape.pipe = st_options('escape.pipe'), …)

Arguments

x
A summarytools object that was generated with freq, descr, ctable or dfSummary.

method
One of “pander”, “viewer”, “browser”, or “render”. For print(), default is “pander”; for view(), default is “viewer”. If “viewer” is used outside RStudio, “browser” will be used instead. Use “render” if function is called from an Rmd document.

file
File name to write output to. Defaults to “”.

append
Logical. When file argument is supplied, this indicates whether to append output to existing file. FALSE by default.

report.title
For html reports, this goes into the <title> tag. Defaults to NA, in which case <title> will be generic.

table.classes
Character. Additional classes to assign to output tables. All Bootstrap CSS classes can be used. It also allows user-defined classes (see custom.css parameter). See details section. NA by default.

bootstrap.css
Logical. Set to FALSE to omit Bootstap css. TRUE by default. To change this default value globally, see st_options.

custom.css
Path to a user-defined .css file. Classes defined in this file can be used in the table.classes parameter. NA by default. To change this default value globally, see st_options.

silent
Hide console messages (such as ignored variables or NaN to NA transformations).

footnote
footnote in html output. When set to “default”, this is the package name and version, R version, and current date). Has no effect when method is “pander”. Set to “default”, provide your own text, or set to NA to omit. To change this default value globally, see st_options.

escape.pipe
Logical. Set to TRUE when using style='grid' and file argument is supplied if the intent is to generate a text file that can be converted to other formats using Pandoc. To change this default value globally, see st_options.

…
Additional arguments can be used to override parameters stored as attributes in the object being printed. See Details section.

Value

NULL when method="pander"; a file path (returned invisibly) when method="viewer" or method="browser". In the latter case, the file path is also passed to shell.exec so the document is opened with default Web Browser.

Details

Plain ascii and rmarkdown tables are generated via pander. See References section for a list of all available pander options.

To print objects of class “by”, use view. This function also makes it more practical to generate html files (see examples).

The following additional arguments can be used to override formatting and other attributes stored in the object to be printed. Refer to the function's documentation for details on these arguments.

style

round.digits (except for dfSummary objects)

justify

plain.ascii

missing

Data.type

Subset

Group

Weights

date

omit.headings

split.tables

Dataframe

Dataframe.label

Variable

Variable.label

display.labels

display.type

totals (freq and ctable objects)

report.nas (freq objects only)

Row.variable (ctable objects only)

Col.variable (ctable objects only)

Row.variable.subset (ctable objects only)

Col.variable.subset (ctable objects only)

Row.variable.label (ctable objects only)

Col.variable.label (ctable objects only)

References

Rstudio Summarytools on Github List of pander options on Github Bootstrap Cascading Stylesheets

See also

Examples

  
# NOT RUN {
  data(tobacco)
  view(dfSummary(tobacco), footnote = NA)

# }
  data(exams)
  print(freq(exams$gender), style = 'rmarkdown')

#> ### Frequencies   
#> **Variable:** exams$gender     
#> **Type:** Factor (unordered)   
#> 
#> |     &nbsp; | Freq | % Valid | % Valid Cum. | % Total | % Total Cum. |
#> |-----------:|-----:|--------:|-------------:|--------:|-------------:|
#> |   **Girl** |   15 |   50.00 |        50.00 |   50.00 |        50.00 |
#> |    **Boy** |   15 |   50.00 |       100.00 |   50.00 |       100.00 |
#> | **\<NA\>** |    0 |         |              |    0.00 |       100.00 |
#> |  **Total** |   30 |  100.00 |       100.00 |  100.00 |       100.00 |

  print(descr(exams), omit.headings = TRUE)

#> Non-numerical variable(s) ignored: student, gender

#> 
#>                     french    math   geography   history   economics   english
#> ----------------- -------- ------- ----------- --------- ----------- ---------
#>              Mean    73.94   73.54       70.04     72.77       73.91     75.96
#>           Std.Dev    10.79    9.19       10.65     10.20        8.62      7.92
#>               Min    44.80   55.60       47.20     53.90       60.50     58.30
#>                Q1    68.20   66.95       65.90     68.20       68.80     70.90
#>            Median    73.60   73.75       68.50     72.75       71.60     74.10
#>                Q3    76.70   80.35       77.80     76.50       77.00     80.60
#>               Max    94.70   93.20       96.30     93.50       94.20     93.10
#>               MAD     7.56    9.93       12.31      6.45        5.49      6.52
#>               IQR     8.50   13.35       11.90      8.15        8.20      9.70
#>                CV     0.15    0.12        0.15      0.14        0.12      0.10
#>          Skewness     0.03    0.12        0.10      0.01        0.75      0.28
#>       SE.Skewness     0.43    0.44        0.43      0.43        0.43      0.43
#>          Kurtosis     0.45   -0.58       -0.03     -0.60       -0.42     -0.25
#>           N.Valid    29.00   28.00       29.00     30.00       29.00     29.00
#>         Pct.Valid    96.67   93.33       96.67    100.00       96.67     96.67