Displaying and setting summarytools global options

To list all summarytools global options, run this function without any parameters. To display the value of an option, use the first parameter only. To modify it, add the new value as a second parameter.

st_options(option, value)

Arguments

option
option name (string).

value
value to assign (optional)

Details

The following options are available:

style Character. One of “simple” (default), “rmarkdown”, or “grid”.

plain.ascii Logical. TRUE by default. Set to FALSE when using summarytools with a rendering tool such as knitr or when creating rmarkdown output files to be converted with Pandoc (although note that its value will automatically be set to FALSE whenever style = “rmarkdown”).

round.digits Numeric. Defaults to 2.

omit.headings Logical. Set to TRUE to remove all headings from outputs (only the tables will be printed out). FALSE by default.

footnote Character. When the default value “default” is used, the package name, version, and R version are displayed below html outputs. Set no NA to omit the footnote, or provide a string to personalize it.

display.labels Logical. TRUE by default. Set to FALSE to omit data frame and variable labels in the headings section.

freq.totals Logical. Corresponds to the totals parameter of freq. TRUE by default.

freq.display.nas Logical. Corresponds to the display.nas parameter of freq(). TRUE by default.

ctable.totals Logical. Corresponds to the totals parameter of ctable. TRUE by default.

ctable.prop Character. Corresponds to the prop parameter of ctable. Defaults to “r” (row).

descr.stats Character. Corresponds to the stats parameter of descr. Defaults to “all”.

descr.transpose Logical. Corresponds to the transpose parameter of descr. FALSE by default.

bootstrap.css Logical. Include Bootstrap CSS in html outputs. Defaults to TRUE. Set to FALSE When using the “render” method inside a shiny app.

custom.css Character. Path to an additional, user-provided, CSS file. NA by default.

escape.pipe Logical. Set to TRUE if Pandoc conversion is your goal and you have unsatisfying results with grid or multiline tables. FALSE by default.

Note

Loosely based on Gergely Daróczi's panderOptions function.

Examples

# NOT RUN {
st_options()                   # show all summarytools global options
st_options('round.digits')     # show a specific global option 
st_options('round.digits', 1)  # set an option
st_options('reset')            # reset all summarytools global options
# }