stby {summarytools}	R Documentation

Obtain Grouped Statistics With summarytools

Description

This is essentially the base by function, except for the class of the returned object.

Usage

stby(data, INDICES, FUN, ..., simplify = TRUE)

Arguments

data
an R object, normally a data frame, possibly a matrix.

INDICES
a grouping variable or a list of grouping variables, each of length nrow(data).

FUN
a function to be applied to (usually data-frame) subsets of data.

...
Further arguments to FUN.

simplify
Logical. Essentially a placeholder to maintain full compatibility with base by. For more details, see tapply.

Value

An object having classes “list” and “summarytools”.

See Also

by, tapply

Examples

data("tobacco")
with(tobacco, stby(BMI, gender, descr))

[Package summarytools version 1.0.1 Index]