summarytools-package {summarytools}	R Documentation

Tools to Quickly and Neatly Summarize Data

Description

summarytools is a collection of functions which neatly and quickly summarize numerical and categorical data. Data frame summaries, frequency tables and cross-tabulations, as well as common descriptive (univariate) statistics can be produced in a straightforward manner. Users with little to no prior R programming experience but who are familiar with popular commercial statistical software such as SAS, SPSS and Stata will feel right at home.

Details

These are the four core functions:

dfSummary

Extensive yet legible data frame summaries.

freq

Frequency tables supporting weights and displaying proportions of valid and of total data, including cumulative proportions.

descr

All common univariate descriptive stats applied to a single vector or to all numerical vectors contained in a data frame.

ctable

Cross-tabulations for pairs of categorical variables – accepting both numerical and character vectors, as well as factors. Choose between Total, Columns or Rows proportions, and optionally display chi-square statistic (with corresponding p-value), odds ratio, as well as risk ratio with flexible confidence intervals.

Choice of output formats:

plain ascii

Ideal when showing results in the R console.

rmarkdown

Perfect for writing short papers or presentations.

html

A format very well integrated in RStudio – but will work with any Web browser. Use the view function to display results directly in RStudio's viewer, or in your preferred Web browser.

Author(s)

Maintainer: Dominic Comtois dominic.comtois@gmail.com

See Also

Useful links:

https://github.com/dcomtois/summarytools

Report bugs at https://github.com/dcomtois/summarytools/issues

[Package summarytools version 1.0.1 Index]