use_custom_lang {summarytools}	R Documentation

Import and use a custom language

Description

If your language is not available or if you wish to customize the outputs' language to suit your preference, you can set up a translations file (see details) and import it with this function.

Usage

use_custom_lang(file)

Arguments

file
Character. The path to the translations file.

Details

To build the translations file, copy the language_template.csv file located in the installed package's includes directory and fill out the ‘custom’ column using a text editor, leaving column titles unchanged. The file must also retain its UTF-8 encoding.

[Package summarytools version 1.0.1 Index]