view {summarytools}	R Documentation

view

Description

Visualize results in RStudio's Viewer or in Web Browser

Usage

view(x, method = "viewer", file = "", append = FALSE,
  report.title = NA, table.classes = NA, 
  bootstrap.css = st_options("bootstrap.css"), 
  custom.css = st_options("custom.css"), silent = FALSE, 
  footnote = st_options("footnote"),
  max.tbl.height = Inf,
  collapse = 0,
  escape.pipe = st_options("escape.pipe"), ...)

Arguments

x
A summarytools object, created by one of the four core functions (freq, descr, ctable, or dfSummary).

method
Character. One of “pander”, “viewer”, “browser”, or “render”. Default value for the print() method is “pander”; for view()/stview(), default is “viewer” if session is running in RStudio, “browser” otherwise. The main use for “render” is in R Markdown documents.

file
Character. File name to write output to. Defaults to “”.

append
Logical. Append output to existing file (specified using the file argument). FALSE by default.

report.title
Character. For html reports, this goes into the <title> tag. When left to NA (default), the first line of the heading section is used (e.g.: “Data Frame Summary”).

table.classes
Character. Additional html classes to assign to output tables. Bootstrap css classes can be used. User-defined classes (see the custom.css argument) are also specified here. See details section. NA by default.

bootstrap.css
Logical. When generating an html document, include the “includes/stylesheets/bootstrap.min.css"” file content inside a <style type="text/css"> tag in the document's <head>. TRUE by default. Can be set globally with st_options.

custom.css
Character. Path to a custom .css file. Classes defined in this must also appear in the table.classes parameter in order to be applied to the table(s). Can be set globally with st_options. NA by default.

silent
Logical. Set to TRUE to hide console messages (e.g.: ignored variables or NaN to NA transformations). FALSE by default.

footnote
Character. Text to display just after html output tables. The default value (“default”) produces a two-line footnote indicating the package's name and version, the R version, and the current date. Has no effect on ascii or markdown content. Can contain standard html tags. Set to NA to omit. Can be set globally with st_options.

max.tbl.height
Numeric. Maximum table height in pixels allowed in rendered dfSummary() tables. When this argument is used, results will show up in a <div> with the specified height and a scroll bar. Intended to be used in Rmd documents with method = "render". Inf by default.

collapse
Numeric. 0 by default. Set to 1 to make freq() sections collapsible (when clicking on the variable name). Future versions might provide alternate collapsing options.

escape.pipe
Logical. Set to TRUE when style="grid" and file argument is supplied if the intent is to generate a text file that can be converted to other formats using Pandoc. Can be set globally with st_options.

...
Additional arguments used to override attributes stored in the object, or to change formatting via format or pander. See Details.

Details

Creates html outputs and displays them in RStudio's viewer, in a browser, or renders the html code in R markdown documents.

For objects of class “summarytools”, this function is simply a wrapper around print.summarytools with method = "viewer".

Objects of class “by”, “stby”, or “list” are dispatched to the present function, as it can manage multiple objects, whereas print.summarytools can only manage one object at a time.

[Package summarytools version 1.0.1 Index]