parse_args {summarytools}	R Documentation

Extract Data Information From Arguments Passed to Functions

Description

Using sys.calls(), sys.frames() and match.call(), this utility function extracts and/or infers information about the data being processed. Data frame name, variable names and labels if any, subsetting information, grouping information (when by() is used) are returned by the function which tries various methods to get this information.

Usage

parse_args(
  sys_calls,
  sys_frames,
  match_call,
  var = "x",
  silent = FALSE,
  df_name = TRUE,
  df_label = TRUE,
  var_name = TRUE,
  var_label = TRUE,
  caller = ""
)

Arguments

sys_calls
Object created using sys.calls().

sys_frames
Object created using sys.frames().

match_call
Object created using match.call().

var
Character. “x” (default) and/or “y” (the latter being used only in ctable).

silent
Logical. Hide console messages. TRUE by default.

Value

A list consisting of one or many of the following items

df_name The data frame name

df_label The data frame label

var_name The variable name(s)

var_label The variable label

by_var The variable used in by(), when in the call stack

by_group The group, when by() was used

by_first Binary indicator used when by() is in the call stack

by_last Binary indicator

Author(s)

Dominic Comtois, dominic.comtois@gmail.com

[Package summarytools version 1.0.1 Index]