what.is {summarytools}	R Documentation

Obtain Extended Properties of Objects

Description

Combination of most common “macro-level” functions that describe an object.

Usage

what.is(x, ...)

Arguments

x
Any object.

...
Included for backward-compatibility only. Has no real use.

Details

An alternative to calling in turn class, typeof, dim, and so on. A call to this function will readily give all this information at once.

Value

A list with following elements:

properties

A data frame with the class(es), type, mode and storage mode of the object as well as the dim, length and object.size.

attributes.lengths

A named character vector giving all attributes (c.f. “names”, “row.names”, “class”, “dim”, and so forth) along with their length.

extensive.is

A character vector of all the identifier functions. (starting with “is.”) that yield TRUE when used with x as argument.

function.type

When x is a function, results of ftype are added.

Author(s)

Dominic Comtois, dominic.comtois@gmail.com

See Also

class, typeof, mode, storage.mode, dim, length, is.object, otype, object.size, ftype

Examples

what.is(1)
what.is(NaN)
what.is(iris3)
what.is(print)
what.is(what.is)

[Package summarytools version 1.0.1 Index]